refactoring entitiy
