+++ week4/demo/src/main/java/com/example/demo/domain/member/Food.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.member;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
