+++ week4/demo/src/main/java/com/example/demo/domain/mission/Mission.java
@@ -1,5 +1,6 @@
-package com.example.demo.domain;
+package com.example.demo.domain.mission;
 
+import com.example.demo.domain.restruant.FoodMarket;
 import jakarta.persistence.*;
 import lombok.*;
