+++ week4/demo/src/main/java/com/example/demo/domain/review/ReviewImage.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.review;
 
 import jakarta.persistence.*;
 import lombok.*;
